+++ Sources/AudioSTT/AudioSTT.swift
@@ -1,2 +1,5 @@
+@_exported import AudioQwen3ASRModel
+@_exported import AudioSortformer
+
 /// AudioSTT module marker.
 public enum AudioSTTModule {}